+++ src/ui.rs
@@ -14,6 +14,7 @@ mod dialog_encryption_password;
 mod dialog_exclude;
 mod dialog_exclude_pattern;
 mod dialog_info;
+mod dialog_preferences;
 mod dialog_prune;
 mod dialog_prune_review;
 mod dialog_setup;
@@ -140,6 +141,7 @@ fn on_startup(_app: &adw::Application) {
     adw_app().set_accels_for_action("app.help", &["F1"]);
     adw_app().set_accels_for_action("app.quit", &["<Ctrl>Q"]);
     adw_app().set_accels_for_action("app.setup", &["<Ctrl>N"]);
+    adw_app().set_accels_for_action("app.backup-preferences", &["<Ctrl>comma"]);
     adw_app().set_accels_for_action("win.show-help-overlay", &["<Ctrl>question"]);
 
     if BACKUP_CONFIG.load().iter().count() == 1 {