+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.7.0 (unreleased)
 
 - Add docker to containers and VMs exclusion preset
+- Add per-backup preferences window
 
 ## 0.6.0 (2023-04-13)